Adjusted the wordpress script

+++ src/requirements/wordpress/script.sh
@@ -1,9 +1,32 @@
 #!/bin/sh
 cd /var/www/html
-curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
-chmod +x wp-cli.phar
-php84 -d memory_limit=-1 wp-cli.phar core download --allow-root
-php84 -d memory_limit=-1 wp-cli.phar config create --dbname=wordpress --dbuser=wpuser --dbpass=password --dbhost=mariadb --allow-root
-php84 -d memory_limit=-1 wp-cli.phar core install --url=localhost --title=inception --admin_user=admin --admin_password=admin --admin_email=admin@admin.com --allow-root
 
-php-fpm84 -F
+if [ ! -f wp-config.php ]; then
+    curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
+    chmod +x wp-cli.phar
+
+    php84 -d memory_limit=-1 wp-cli.phar core download --allow-root
+
+    php84 -d memory_limit=-1 wp-cli.phar config create \
+        --dbname=$MYSQL_DATABASE \
+        --dbuser=$MYSQL_USER \
+        --dbpass=$MYSQL_PASSWORD \
+        --dbhost=mariadb \
+        --allow-root
+
+    php84 -d memory_limit=-1 wp-cli.phar core install \
+        --url=$DOMAIN_NAME \
+        --title=inception \
+        --admin_user=$WP_ADMIN_USER \
+        --admin_password=$WP_ADMIN_PASSWORD \
+        --admin_email=$WP_ADMIN_EMAIL \
+        --allow-root
+
+    php84 -d memory_limit=-1 wp-cli.phar user create \
+        $WP_USER $WP_USER_EMAIL \
+        --role=author \
+        --user_pass=$WP_USER_PASSWORD \
+        --allow-root
+fi
+
+exec php-fpm84 -F
